chore: move IUVlist component - move consts files in utils/consts - remove todo - remove old changeset
