--- .changeset/cold-sheep-hunt.md
@@ -1,6 +1,0 @@
----
-"io-ipatente-payments": minor
-"@io-ipatente/ui": patch
----
-
-Init summary page for payments - listItemAction icon become optional
